Rewrote StreamSummary to consume a lot less memory. However, it's 2x-6x slower than list-based version.

+++ src/main/java/com/clearspring/analytics/stream/ConcurrentStreamSummary.java
@@ -60,7 +60,7 @@ public class ConcurrentStreamSummary<T> implements ITopK<T>
 	}
 
 	@Override
-	public boolean offer(final T element, final int incrementCount)
+	public boolean offer(final T element, final long incrementCount)
 	{
 		long val = incrementCount;
 		ScoredItem value = new ScoredItem(element, incrementCount);
